fix(lapack): skip MAX.3/MIN.3 rename when files are absent
The manpages tarball for lapack 3.9.0 does not include MAX.3 or
MIN.3, so the unconditional mv (added for rhbz#2222868) fails.
Make the rename conditional so the build succeeds regardless.

+++ root/ppg/common/deps/lapack/rpm/lapack.spec
@@ -438,9 +438,9 @@ popd
 
 #rhbz#2222868
 pushd manpages/man/man3
-mv MAX.3 lapack-MAX.3
-mv MIN.3 lapack-MIN.3
-popd 
+[ -f MAX.3 ] && mv MAX.3 lapack-MAX.3 || :
+[ -f MIN.3 ] && mv MIN.3 lapack-MIN.3 || :
+popd
 
 find manpages/man/man3 -type f -printf "%{_mandir}/man3/%f*\n" > lapackmans
